Refactor :  메인으로 가는 버튼 추가
